refactor: useCategoryInfo -> improve hook to include more category info data needed for PackGraphic
