+++ src/components/Dashboard/PackInfo/PackGraphic.tsx
@@ -2,34 +2,24 @@ import { Category } from '../../../types/packTypes';
 import PackChart from '../PackChart/PackChart';
 import { Image, Icon, List, ListItem, Label, Divider } from 'semantic-ui-react';
 import CampGraphic from '../../../assets/camping.svg';
-import { useWeightSum } from './useWeightSum';
+import { useCategoryInfo } from './useCategoryInfo';
 
 type PackGraphicProps = {
 	fetching: boolean;
 	packCategories: Category[];
 };
 
-const chartColors = ['#338866', '#78B87A', '#5F84A2', '#7BB8C0', '#A7B5FE', '#F36F3B'];
-
 const PackGraphic = (props: PackGraphicProps) => {
 	const { fetching, packCategories } = props;
-	const { categoryWeights, packHasWeight } = useWeightSum(packCategories);
-
-	const totalWeight = categoryWeights.reduce((curr, acc) => (acc += curr), 0);
-
-	const displayList = categoryWeights.map((totalWeight: number, index) => {
-		const categoryName = packCategories[index].packCategoryName;
-		const categoryId = packCategories[index].packCategoryId;
-		const chartColor = chartColors[index];
-		return { categoryName, categoryId, totalWeight, chartColor };
-	});
+	const { chartCategoryInfo, categoryWeights, totalWeight, packHasWeight } =
+		useCategoryInfo(packCategories);
 
 	if (packHasWeight && !fetching) {
 		return (
 			<>
 				<div className="pack-info-center-panel">
 					<List className="chart-display-list">
-						{displayList.map((category) => {
+						{chartCategoryInfo.map((category) => {
 							return (
 								<ListItem key={category.categoryId} className="chart-display-list-item">
 									<Icon name="circle" style={{ color: category.chartColor || 'grey' }} />
